AddToCart

+++ src/main/java/asm/java5Nhom6/entity/Color.java
@@ -20,7 +20,7 @@ public class Color {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "color_id")
-    private Long colorId;
+    private Integer colorId;
 
     @Column(name = "name", nullable = false)
     private String name;
